backend/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
django==4.0
gunicorn==20.1.0
python-dotenv==0.21.0
google-api-python-client==2.64.0
google_auth_oauthlib
django-cors-headers
sendgrid
django-environ
